Migration PostgreSQL : load.py, dbt (marts corrigés), ajout du dossier airflow (DAG + docker-compose)

+++ dbt_agro_meteo/models/marts/mart_rendement_par_region.sql
@@ -5,8 +5,8 @@ select
     region,
     culture,
     annee_recolte,
-    round(avg(rendement), 3) as rendement_moyen,
-    round(sum(surface_ha), 1) as surface_totale_ha,
-    round(sum(production_tonnes), 1) as production_totale_tonnes
+    round(avg(rendement)::numeric, 3) as rendement_moyen,
+    round(sum(surface_ha)::numeric, 1) as surface_totale_ha,
+    round(sum(production_tonnes)::numeric, 1) as production_totale_tonnes
 from {{ ref('stg_production_agricole') }}
 group by pays, region, culture, annee_recolte
